+++ docs/app/layout.tsx
@@ -4,6 +4,7 @@ import { Inter } from "next/font/google";
 import type { ReactNode } from "react";
 import { ProvidersWrapper } from "@/lib/providers/providers-wrapper";
 import { Banners } from "@/components/layout/banners";
+import Script from "next/script";
 
 const inter = Inter({
   subsets: ["latin"],
@@ -12,10 +13,21 @@ const inter = Inter({
 export default async function Layout({ children }: { children: ReactNode }) {
   return (
     <html lang="en" className={inter.className} suppressHydrationWarning>
+      <head>
+        <Script
+          id="hubspot-script"
+          type="text/javascript"
+          src="//js.hs-scripts.com/45532593.js"
+          async
+          defer
+        />
+      </head>
       <body>
         <ProvidersWrapper>
           <Banners />
-          <RootProvider theme={{ enabled: true, defaultTheme: 'dark' }}>{children}</RootProvider>
+          <RootProvider theme={{ enabled: true, defaultTheme: "dark" }}>
+            {children}
+          </RootProvider>
         </ProvidersWrapper>
       </body>
     </html>