+++ CopilotKit/packages/react-textarea/CHANGELOG.md
@@ -1,5 +1,22 @@
 # ui
 
+## 1.8.12-next.6
+
+### Patch Changes
+
+- Updated dependencies [3e09584]
+  - @copilotkit/react-core@1.8.12-next.6
+  - @copilotkit/runtime-client-gql@1.8.12-next.6
+  - @copilotkit/shared@1.8.12-next.6
+
+## 1.8.12-next.5
+
+### Patch Changes
+
+- @copilotkit/react-core@1.8.12-next.5
+- @copilotkit/runtime-client-gql@1.8.12-next.5
+- @copilotkit/shared@1.8.12-next.5
+
 ## 1.8.12-next.4
 
 ### Patch Changes