+++ CopilotKit/packages/react-ui/CHANGELOG.md
@@ -1,5 +1,23 @@
 # ui
 
+## 1.8.12-next.6
+
+### Patch Changes
+
+- Updated dependencies [3e09584]
+  - @copilotkit/react-core@1.8.12-next.6
+  - @copilotkit/runtime-client-gql@1.8.12-next.6
+  - @copilotkit/shared@1.8.12-next.6
+
+## 1.8.12-next.5
+
+### Patch Changes
+
+- 9b2e9e6: - fix: move powered by tag to input and fix padding on full chat
+  - @copilotkit/react-core@1.8.12-next.5
+  - @copilotkit/runtime-client-gql@1.8.12-next.5
+  - @copilotkit/shared@1.8.12-next.5
+
 ## 1.8.12-next.4
 
 ### Patch Changes