+++ CopilotKit/packages/shared/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @copilotkit/shared
 
+## 1.8.12-next.6
+
+## 1.8.12-next.5
+
 ## 1.8.12-next.4
 
 ## 1.8.12-next.3